backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
qdrant-client
python-dotenv
email-validator
pydantic[email]
google-generativeai
fastembed